Change test_import to use rstest fixture as well.
