+++ cli/tests/testdata/import/csv_multi_currency.csv
