+++ cli/tests/testdata/import/index_amount.csv
