+++ core/Cargo.toml
@@ -30,7 +30,7 @@ env_logger.workspace = true
 indoc.workspace = true
 maplit.workspace = true
 pretty_assertions.workspace = true
-rstest = "0.24"
+rstest.workspace = true
 rust_decimal_macros.workspace = true
 criterion = {version = "0.5", features = ["html_reports"]}
